Refactored User data classes into User.data
