+++ auth-service/src/main/java/com/petclinic/auth/User/UserRepo.java
@@ -12,6 +12,7 @@
  */
 package com.petclinic.auth.User;
 
+import com.petclinic.auth.User.data.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
