+++ auth-service/src/main/java/com/petclinic/auth/User/UserServiceImpl.java
@@ -3,6 +3,10 @@ import com.petclinic.auth.Exceptions.IncorrectPasswordException;
 import com.petclinic.auth.JWT.JWTService;
 import com.petclinic.auth.Mail.Mail;
 import com.petclinic.auth.Mail.MailService;
+import com.petclinic.auth.User.data.User;
+import com.petclinic.auth.User.data.UserIDLessRoleLessDTO;
+import com.petclinic.auth.User.data.UserPasswordLessDTO;
+import com.petclinic.auth.User.data.UserTokenPair;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
