+++ auth-service/src/main/java/com/petclinic/auth/User/data/UserPasswordLessDTO.java
@@ -5,7 +5,7 @@
  * Date: 09/10/21
  * Ticket: feat(AUTH-CPC-310)
  */
-package com.petclinic.auth.User;
+package com.petclinic.auth.User.data;
 import com.petclinic.auth.Role.Role;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
