+++ auth-service/src/test/java/com/petclinic/auth/AuthServicePersistenceTests.java
@@ -22,7 +22,7 @@ package com.petclinic.auth;
 
 import com.petclinic.auth.Role.Role;
 import com.petclinic.auth.Role.RoleRepo;
-import com.petclinic.auth.User.User;
+import com.petclinic.auth.User.data.User;
 import com.petclinic.auth.User.UserRepo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
