+++ auth-service/src/test/java/com/petclinic/auth/JWT/JWTServiceTests.java
@@ -1,7 +1,7 @@
 package com.petclinic.auth.JWT;
 
 import com.petclinic.auth.Role.Role;
-import com.petclinic.auth.User.User;
+import com.petclinic.auth.User.data.User;
 import io.jsonwebtoken.JwtException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
