+++ auth-service/src/test/java/com/petclinic/auth/User/AuthServiceUserEntityTests.java
@@ -1,6 +1,7 @@
 package com.petclinic.auth.User;
 
 import com.petclinic.auth.Role.Role;
+import com.petclinic.auth.User.data.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
