+++ auth-service/src/test/java/com/petclinic/auth/User/AuthServiceUserRepoTests.java
@@ -13,6 +13,7 @@
  */
 package com.petclinic.auth.User;
 
+import com.petclinic.auth.User.data.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
